pyproject.toml
[tool.poetry]
name = "sertest"
version = "0.1.0"
description = ""
authors = ["Chen Yang <yangchen@thinkmore.info>"]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
